requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.60.0
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
requests==2.32.3
